chore: refactor on some logic for component manifests resources (#1121)
* chore: refactor on some logic for component manifests resources

* update: move functions to other files to make deploy.so neat


---------
